model_docker/requirements.txt
wheel
attrs==21.2.0
Automat==20.2.0
azure-core==1.20.1
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
characteristic==14.3.0
charset-normalizer==2.0.7
constantly==15.1.0
cryptography==3.4.7
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
isodate==0.6.0
klein==20.6.0
msrest==0.6.21
numpy==1.21.4
oauthlib==3.1.1
pandas==1.3.4
psycopg2==2.7.7
pycparser==2.21
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
Tubes==0.2.0
Twisted==21.7.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
zope.interface==5.4.0
joblib==1.0.1
catboost
azure-keyvault-secrets
azure-identity